requirements.txt
python-telegram-bot==20.7
Pillow
reportlab
PyPDF2